Removed the class PixelColor and inlined the single function still used.
This class looked a good ideat at the beginning, when I was thinking
to send the data pixel by pixel to keep the android API easier to
read.

But also android api handles colors packed in a integer. Therefore
this class quickly become just a way to convert this information.
And it was pushed lower and lower into the stack, to the point that it
was used in the class that already handles all the bytes.